+++ crates/holochain_conductor_api/CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - **BREAKING CHANGE**: Rename `AppInfo` to `GetAppInfo`.
 - **BREAKING CHANGE**: Rename `ZomeCall` to `CallZome`. [\#1707](https://github.com/holochain/holochain/pull/1707)
 
+- **BREAKING CHANGE**: Rename ArchiveCloneCell to DisableCloneCell.
+- **BREAKING CHANGE**: Rename RestoreArchivedCloneCell to EnableCloneCell.
+- **BREAKING CHANGE**: Move EnableCloneCell to App API.
+- **BREAKING CHANGE**: Refactor DeleteCloneCell to delete a single disabled clone cell. [\#1704](https://github.com/holochain/holochain/pull/1704)
+
 ## 0.0.72
 
 ## 0.0.71