+++ crates/holochain_conductor_api/CHANGELOG.md
@@ -7,17 +7,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+## 0.1.0-beta-rc.0
+
 - **BREAKING CHANGE**: Remove deprecated Admin and App API calls.
+
 - **BREAKING CHANGE**: Remove call `InstallApp`.
+
 - **BREAKING CHANGE**: Rename `InstallAppBundle` to `InstallApp`.
+
 - **BREAKING CHANGE**: Rename `ZomeCall` to `CallZome`. [\#1707](https://github.com/holochain/holochain/pull/1707)
 
 - **BREAKING CHANGE**: Rename ArchiveCloneCell to DisableCloneCell.
+
 - **BREAKING CHANGE**: Rename RestoreArchivedCloneCell to EnableCloneCell.
+
 - **BREAKING CHANGE**: Move EnableCloneCell to App API.
+
 - **BREAKING CHANGE**: Refactor DeleteCloneCell to delete a single disabled clone cell. [\#1704](https://github.com/holochain/holochain/pull/1704)
 
 - **BREAKING CHANGE**: Refactor `AppInfo` to return all cells and DNA modifiers.
+
 - **BREAKING CHANGE**: Rename `RequestAgentInfo` to `AgentInfo`. [\#1719](https://github.com/holochain/holochain/pull/1719)
 
 ## 0.0.72